pyproject.toml
[tool.poetry]
name = "markit"
version = "0.1.0"
description = ""
authors = ["Antonio Fernando <antoniofernandodj@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
dynaconf = "^3.2.6"
pymysql = "^1.1.1"
cryptography = "^43.0.1"
pydantic = {extras = ["email"], version = "^2.9.2"}
aiomysql = "^0.2.0"
bcrypt = "^4.2.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
uvicorn = "^0.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.6.8"
pytest-twisted = "^1.14.3"
twisted = "^24.7.0"
pytest-asyncio = "^0.24.0"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
